requirements.txt
# Core Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# AI/ML Framework
torch==2.0.1
torchvision==0.15.2

# Computer Vision
opencv-python==4.8.0.74
Pillow==10.0.0

# Numerical Computing
numpy==1.24.3

# Production (Optional)
gunicorn==20.1.0
python-dotenv==1.0.0